backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiofiles>=23.2.1
pydantic>=2.7.0
python-multipart>=0.0.9
websockets>=12.0
pyyaml>=6.0.1
